Fix reverse debugging when going back into functions

+++ yp-gdb.py
@@ -5,7 +5,9 @@ import functools
 import contextlib
 
 log = logging.getLogger(__name__)
-log.setLevel(logging.INFO)
+#log.setLevel(logging.INFO)
+log.setLevel(logging.DEBUG)
+log.debug = print
 
 INIT_FRAME_LINE = '../Python/ceval.c:1324'
 END_FRAME_LINE = '../Python/ceval.c:3844'
@@ -86,19 +88,30 @@ def invoke(cmd, args, from_tty):
 
 class StackFrame:
     def __init__(self):
-        code = gdb.parse_and_eval('co')
-        first_instr = int(gdb.parse_and_eval('first_instr'))
-        filename = gdb.parse_and_eval('co.co_filename')
-        func = gdb.parse_and_eval('co.co_name')
+        gdb.execute('up')  # f might be optimized out, but it is still there in the parent frame
 
-        self.code = code
-        self.filename = filename
-        self.func = func
-        self.first_instr = first_instr
-        self.next_instr = first_instr
-        self.instr_lb = 0
-        self.instr_ub = -1
-        self.instr_prev = -1
+        self.code = int(gdb.parse_and_eval('f.f_code'))
+        lasti = int(gdb.parse_and_eval('f.f_lasti'))
+        
+        q = f'(((PyCodeObject *)({self.code}))->co_code)'
+        q = f'(((PyBytesObject *)({q}))->ob_sval)'
+        q = f'((_Py_CODEUNIT *) {q})'
+        self.first_instr = int(gdb.parse_and_eval(q))
+        self.filename = str(gdb.parse_and_eval('f.f_code.co_filename'))
+        self.func = str(gdb.parse_and_eval('f.f_code.co_name'))
+
+        gdb.execute('down')
+
+        if get_exec_direction() == 'forward':
+            self.next_instr = self.first_instr
+            self.instr_lb = lasti
+            self.instr_ub = lasti
+            self.instr_prev = lasti
+        else:
+            self.next_instr = self.first_instr + lasti
+            self.instr_lb = 0
+            self.instr_ub = -1
+            self.instr_prev = -1
 
     @property
     def lasti(self):
@@ -108,14 +121,18 @@ class StackFrame:
         # If the last instruction falls at the start of a line or if
         # it represents a jump backwards, update the frame's line
         # number and call the trace function.
+
+        # TODO: adjust for reverse debugging
         return lasti == self.instr_lb or lasti < self.instr_prev
 
     @property
     def lineno(self):
         # If the last instruction executed isn't in the current
         # instruction window, reset the window.
-        if self.lasti < self.instr_lb or self.lasti >= self.instr_ub:
-            self.update_lineno()
+
+        # TODO: adjust for reverse debugging
+#        if self.lasti < self.instr_lb or self.lasti >= self.instr_ub:
+        self.update_lineno()
         return self._lineno
 
     def update_lineno(self):
@@ -130,7 +147,7 @@ class StackFrame:
 
 
     def __repr__(self):
-        return f'StackFrame<{self.filename}#{self.lineno} {self.func} lasti={self.lasti}>'
+        return f'StackFrame<{self.filename}#{self.lineno} {self.func} lasti={self.lasti} nexti={self.next_instr:#x} firsti={self.first_instr:#x}>'
 
     def update_instr(self):
         next_instr_expr = gdb.parse_and_eval('next_instr')
@@ -145,28 +162,72 @@ class StackFrame:
             warnings.warn(f'next_instr is optimized out in {fn}#{fn_lineno}', NextInstrOptimizationWarning)
 
 
+def gdb_breakpoint(func_or_line, cls=gdb.Breakpoint):
+    def build_class(func):
+        klass = type(func.__name__, (cls, ), dict(stop=func))
+        return klass
+
+    if callable(func_or_line):
+        func = func_or_line
+        klass = build_class(func)
+        return klass
+    else:
+        line = func_or_line
+        def wrapper2(func):
+            klass = build_class(func)
+            return functools.partial(klass, line)
+        return wrapper2
+
+
 class InitStackFrameBreakpoint(gdb.Breakpoint):
+    def __init__(self, brk_list, *args):
+        print(INIT_FRAME_LINE, *args)
+        gdb.Breakpoint.__init__(self, INIT_FRAME_LINE, *args)
+        self.brk_list = brk_list
+
     def stop(self):
         log.debug(f'stop InitStackFrameBreakpoint depth={len(STACK)} {STACK[-1] if len(STACK) else ""}')
-        STACK.append(StackFrame())
+
+        # TODO: replace by change_frame(self, self.brk_list)
+        if get_exec_direction() == 'forward':
+            enter_frame()
+        else:
+            leave_frame(self, self.brk_list)
+        
         return False
+
+
+def enter_frame():
+    STACK.append(StackFrame())
+
+
+def leave_frame(self_brk, brk_list):
+    if STACK:
+        STACK.pop()
+
+    print(f'leave_frame depth={len(STACK)}')
+
+    if not STACK:
+        for brk in brk_list:
+            if brk != self_brk:
+                brk.delete()
+        self_brk.enabled = False  # TODO we cannot delete ourself
 
 
 class EndStackFrameBreakpoint(gdb.Breakpoint):
     def __init__(self, brk_list, *args):
-        gdb.Breakpoint.__init__ (self, *args)
+        gdb.Breakpoint.__init__(self, END_FRAME_LINE, *args)
         self.brk_list = brk_list
 
     def stop(self):
         log.debug(f'stop EndStackFrameBreakpoint depth={len(STACK)}')
         log.info('get_pygdb_selected_frame', get_pygdb_selected_frame().get_pyop())
-        if STACK:
-            STACK.pop()
 
-        if not STACK:
-            for brk in self.brk_list:
-                brk.delete()
-            self.enabled = False  # TODO we cannot delete ourself
+        if get_exec_direction() == 'forward':
+            leave_frame(self, self.brk_list)
+        else:
+            # TODO raise NotImplemented('What to do in reverse direction here?')
+            enter_frame()
         
         return False
 
@@ -197,48 +258,48 @@ CASE_TARGET_LIST = [f'../Python/ceval.c:{lineno}' for lineno in ceval_case_targe
 
 STEP_MODE = None
 
+@gdb_breakpoint
+def NextInstrBreakpoint(brk):
+    global STEP_MODE
+    frame = STACK[-1]
+    ret = False
 
-class NextInstrBreakpoint(gdb.Breakpoint):
-    def stop(self):
-        global STEP_MODE
-        frame = STACK[-1]
-        ret = False
+    lasti = frame.lasti
+    frame.update_instr()
 
-        lasti = frame.lasti
-        frame.update_instr()
-
-        if frame.is_at_new_line(lasti) and STEP_MODE == 'line':
-            gdb.execute('py-bt')
-            STEP_MODE = None
-            ret = True
-        if STEP_MODE == 'op':
-            STEP_MODE = None
-            ret = True
-
-        log.debug(f'stop NextInstrBreakpoint {frame}')
-        return ret
-
-
-
-class UserInitStackFrameBreakpoint(ConditionalBreakpoint):
-    def stop(self):
-        if not self.should_stop():
-            return False
-    
-        STACK.append(StackFrame())
-        log.debug(f'stop UserInitStackFrameBreakpoint depth={len(STACK)} {STACK[-1]}')
-
-        brk_args = [gdb.BP_BREAKPOINT, 0, True]
-
-        if len(STACK) == 1:
-            init_frame_brk = InitStackFrameBreakpoint(INIT_FRAME_LINE, *brk_args)
-            next_instr_brk_list = [NextInstrBreakpoint(position, *brk_args) for position in CASE_TARGET_LIST]
-            cleanup = [init_frame_brk] + next_instr_brk_list
-            end_frame_brk = EndStackFrameBreakpoint(cleanup, END_FRAME_LINE, *brk_args)
-
+    if frame.is_at_new_line(lasti) and STEP_MODE == 'line':
         gdb.execute('py-bt')
+        STEP_MODE = None
+        ret = True
+    if STEP_MODE == 'op':
+        STEP_MODE = None
+        ret = True
 
-        return True
+    log.debug(f'stop NextInstrBreakpoint {frame}')
+    return ret
+
+
+@gdb_breakpoint(INIT_FRAME_LINE, cls=ConditionalBreakpoint)
+def UserInitStackFrameBreakpoint(brk):
+    if not brk.should_stop():
+        return False
+
+    STACK.append(StackFrame())
+    log.debug(f'stop UserInitStackFrameBreakpoint depth={len(STACK)} {STACK[-1]}')
+
+    brk_args = [gdb.BP_BREAKPOINT, 0, True]
+
+    if len(STACK) == 1:
+        cleanup_list = []
+        init_frame_brk = InitStackFrameBreakpoint(cleanup_list, *brk_args)
+        next_instr_brk_list = [NextInstrBreakpoint(position, *brk_args) for position in CASE_TARGET_LIST]
+        end_frame_brk = EndStackFrameBreakpoint(cleanup_list, *brk_args)
+
+        cleanup_list += [init_frame_brk, end_frame_brk] + next_instr_brk_list
+
+    gdb.execute('py-bt')
+
+    return True
 
 
 @gdb_command("py-break", gdb.COMMAND_BREAKPOINTS, gdb.COMPLETE_NONE)
@@ -249,7 +310,7 @@ def invoke(cmd, args, from_tty):
     where = f'$_streq({gdb_escape(filename)}, {py_string("co.co_filename")})'
     where += f' && $_streq({gdb_escape(func)}, {py_string("co.co_name")})'
 
-    brk = UserInitStackFrameBreakpoint(INIT_FRAME_LINE, where)
+    brk = UserInitStackFrameBreakpoint(where)
 
 
 @gdb_command("py-next", gdb.COMMAND_RUNNING, gdb.COMPLETE_NONE)
